One last push today - bugfixes
Removed the "save output" by default.
Fixed error handling to break to the correct loop.
Reorganized the single NIC function to also use the Job system
Added "Hostname" as an object returned.
Lowered the delay on waiting to pick up an LLDP packet. Submit an issue if you'd like a flag added for slower networks.

## TODO: Move the LLDP object builder to be a proper function, to make way for the CDP functionality.