Removed tcpdump dependency
Merged my changes in with the PsDiscoveryProtocol utilities. Purged all need for the potentially unlicensed copy of tcpdump that was bundled in. All prior functionality has been translated to the new native suite, but I can feel the caching/parsing speed is first on the docket to get optimized - especially if this is being ran frequently on many endpoints.